Remove css-scope-inline entirely.
