--- server/static/css/base.css
@@ -1,69 +1,0 @@
-:root {
-  --font-sans: 'Nippo', sans-serif;
-  --font-mono: 'KodeMono', monospace;
-
-  --lime: #CDFF64;
-  --white-70: rgba(255, 255, 255, 0.7);
-}
-
-html {
-  font-size: 16px;
-  font-family: var(--font-sans);
-}
-
-
-/**
-* This is a variable font
-* You can controll variable axes as shown below:
-* font-variation-settings: 'wght' 700.0;
-*
-* available axes:
-
-* 'wght' (range from 200.0 to 700.0)
-
-*/
-@font-face {
-  font-family: 'Nippo';
-  src: url('../fonts/Nippo-Variable.woff2') format('woff2'),
-    url('../fonts/Nippo-Variable.woff') format('woff'),
-    url('../fonts/Nippo-Variable.ttf') format('truetype');
-  font-weight: 200 700;
-  font-display: swap;
-  font-style: normal;
-}
-
-
-@font-face {
-  font-family: 'KodeMono';
-  src: url('../fonts/KodeMono-Medium.ttf') format('truetype');
-  font-weight: 500;
-  font-display: swap;
-  font-style: normal;
-}
-
-
-/** Further reset; why doesn't modern-normalize do this? */
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  margin: 0;
-}
-
-/* All containers behave this way. */
-.container {
-  margin-left: 0.5rem;
-  margin-right: 0.5rem;
-}
-
-
-/** md is 768px */
-@media screen and (min-width: 768px) {
-  .container {
-    max-width: 640px;
-    margin: 0 auto;
-  }
-}
