--- server/vb/components/button.css
@@ -1,20 +1,0 @@
-me {
-  cursor: pointer;
-  transition: opacity 0.2s ease-in-out;
-  text-transform: uppercase;
-  text-decoration: none;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 100%;
-  border: none;
-  text-align: center;
-  letter-spacing: 0.05em;
-  padding: 20px 24px;
-  background-color: var(--bg-color);
-  color: var(--color);
-}
-
-me:hover {
-  opacity: 0.7;
-  transition: opacity 0.2s ease-in-out;
-}
