--- server/vb/components/check_page.css
@@ -1,116 +1,0 @@
-me {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-me main {
-  width: 100%;
-  text-align: center;
-  padding: 0.5rem 0;
-}
-
-me main img {
-  height: 150px;
-  margin-bottom: -1.75rem;
-}
-
-me main p {
-  font-weight: 378;
-  font-size: 20px;
-  line-height: 130%;
-}
-
-me main h2 {
-  font-weight: 500;
-  font-size: 36px;
-  line-height: 120%;
-  text-transform: uppercase;
-}
-
-me .faq {
-  width: 100%;
-  color: white;
-  padding: 2rem 0;
-}
-
-me .button-holder {
-  display: flex;
-  justify-content: center;
-  margin: 1.5rem 0;
-}
-
-me .form {
-  width: 100%;
-  background-color: white;
-  padding: 2rem 0;
-}
-
-me .urgency {
-  flex-direction: column;
-  gap: 1rem;
-}
-
-@media screen and (min-width: 768px) {
-  me main {
-    padding: 2rem 0;
-  }
-
-  me main img {
-    height: 150px;
-    margin: 1.5rem 0;
-  }
-
-  me .urgency {
-    flex-direction: row;
-    gap: 2rem;
-  }
-}
-
-me main {
-  position: relative;
-  color: var(--main-color);
-  background-color: var(--main-bg-color);
-}
-
-me main a {
-  color: var(--main-color);
-  transition: opacity 0.2s;
-}
-
-me main a:hover {
-  opacity: 0.7;
-  transition: opacity 0.2s;
-}
-
-me main .urgency {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-me main .fireworks {
-  pointer-events: none;
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  overflow: hidden;
-}
-
-me main .separate {
-  padding-left: 1rem;
-}
-
-me main img {
-  display: block;
-}
-
-@media screen and (min-width: 768px) {
-  me main .urgency {
-    flex-direction: row;
-  }
-}
