--- server/vb/components/countdown.css
@@ -1,40 +1,0 @@
-me {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding-bottom: 0.5rem;
-}
-
-me p {
-  text-transform: uppercase;
-}
-
-me .countdown {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 24px;
-  font-weight: 500;
-  font-family: var(--font-mono);
-  gap: 4px;
-  height: 34px !important;
-}
-
-me .countdown span {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  width: 27px;
-}
-
-me .countdown span.number {
-  color: var(--number-color);
-  background-color: var(--number-bg-color);
-}
-
-me .countdown span.colon {
-  color: var(--colon-color);
-  background-color: transparent;
-}
