--- server/vb/components/faq.css
@@ -1,37 +1,0 @@
-me {
-  display: flex;
-  flex-direction: column;
-}
-
-me h2 {
-  font-size: 36px;
-  font-weight: 440;
-  line-height: 130%;
-  margin-bottom: 1rem;
-}
-
-me h3 {
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 28px;
-  margin-top: 1rem;
-}
-
-me p {
-  font-weight: 378;
-  font-size: 18px;
-  line-height: 28px;
-  opacity: 0.7;
-}
-
-me a {
-  color: white;
-  cursor: pointer;
-  text-decoration: underline;
-  transition: opacity 0.2s;
-}
-
-me a:hover {
-  opacity: 0.7;
-  transition: opacity 0.2s;
-}
