--- server/vb/components/finish_check_partial.css
@@ -1,11 +1,0 @@
-me {
-  padding-top: 1rem;
-  margin-left: 0;
-}
-
-@media screen and (min-width: 768px) {
-  me {
-    padding-top: 0;
-    margin-left: 1rem;
-  }
-}
