--- server/vb/components/home_page.css
@@ -1,75 +1,0 @@
-me {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background-color: #cdff64;
-  color: black;
-}
-
-me main {
-  width: 100%;
-  text-align: center;
-  padding: 2rem 0;
-}
-
-me main svg {
-  width: 104px;
-  margin: 1.5rem 0;
-}
-
-@media screen and (min-width: 768px) {
-  me main svg {
-    width: 112px;
-  }
-}
-
-me main p {
-  font-weight: 378;
-  font-size: 20px;
-  line-height: 130%;
-}
-
-me main h2 {
-  font-weight: 500;
-  font-size: 28px;
-  line-height: 140%;
-}
-
-@media screen and (min-width: 768px) {
-  me main h2 {
-    font-size: 32px;
-  }
-}
-
-me .button-holder {
-  display: flex;
-  justify-content: center;
-  margin: 1.5rem 0;
-}
-
-me .ongoing {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 2rem;
-  margin: 2rem 0;
-}
-
-me .upcoming {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  gap: 0.5rem;
-  margin: 0.5rem 0;
-}
-
-me .coming-soon {
-  text-transform: uppercase;
-  font-weight: bold;
-  font-size: 20px;
-  line-height: 130%;
-  display: flex;
-  justify-content: center;
-  margin: 1.5rem 0;
-}
