--- server/vb/components/logo_specimen.css
@@ -1,49 +1,0 @@
-me {
-  display: flex;
-  gap: 0.5rem;
-}
-
-me .bubble {
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-  width: 48px;
-  height: 48px;
-  background-color: var(--logo-bg-color);
-}
-
-me .bubble img {
-  display: block;
-  margin: 0 auto;
-  max-width: 80%;
-  max-height: 80%;
-}
-
-me .bg {
-  display: flex;
-  font-weight: 600;
-  align-items: center;
-  justify-content: center;
-  padding-left: 0.5rem;
-  padding-right: 0.5rem;
-  background-color: var(--logo-bg-color);
-  color: var(--logo-bg-text-color);
-}
-
-me .action {
-  cursor: pointer;
-  display: flex;
-  font-weight: 600;
-  align-items: center;
-  justify-content: center;
-  padding-left: 0.5rem;
-  padding-right: 0.5rem;
-  transition: opacity 0.2s;
-  background-color: var(--logo-action-color);
-  color: var(--logo-action-text-color);
-}
-
-me .action:hover {
-  opacity: 0.7;
-  transition: opacity 0.2s ease-in-out;
-}
