--- server/vb/components/rules_page.css
@@ -1,5 +1,0 @@
-me {
-  font-size: 1.25em;
-  line-height: 150%;
-  font-family: ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
-}
