--- server/vb/components/school_page.css
@@ -1,68 +1,0 @@
-me {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-me main {
-  width: 100%;
-  text-align: center;
-  padding-bottom: 2rem;
-  color: var(--color);
-  background-color: var(--bg-color);
-}
-
-me main>div {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  min-height: calc(100dvh - env(safe-area-inset-bottom) - 2rem);
-}
-
-@media screen and (min-width: 768px) {
-  me main>div {
-    padding: 2rem 0;
-    min-height: unset;
-  }
-}
-
-me main img {
-  height: 150px;
-  margin: 1.5rem 0;
-}
-
-me main p {
-  font-weight: 378;
-  font-size: 20px;
-  line-height: 130%;
-}
-
-me main h2 {
-  font-weight: 500;
-  font-size: 36px;
-  line-height: 120%;
-  text-transform: uppercase;
-}
-
-me .faq {
-  width: 100%;
-  color: white;
-  padding: 2rem 0;
-}
-
-me .button-holder {
-  display: flex;
-  justify-content: center;
-  margin: 0;
-}
-
-@media screen and (min-width: 768px) {
-  me .button-holder {
-    margin-top: 1.5rem;
-  }
-}
-
-me .faq {
-  background-color: black;
-}
