--- server/vb/components/upcoming_contest.css
@@ -1,34 +1,0 @@
-me {
-  border: 3px solid black;
-  padding: 1rem;
-  color: black;
-  font-size: 18px;
-  font-weight: 440;
-  font-variation-settings: "wght" 440;
-  line-height: 1;
-}
-
-me .content {
-  display: flex;
-  align-items: center;
-  gap: 1em;
-}
-
-me .logo {
-  border-radius: 100%;
-  border: 2px solid black;
-  background-color: var(--logo-bg-color);
-  overflow: hidden;
-  width: 36px;
-  height: 36px;
-}
-
-me .logo img {
-  width: 100%;
-  height: 100%;
-  object-fit: contain;
-}
-
-me p {
-  margin: 0;
-}
