--- server/vb/components/validate_email_page.css
@@ -1,110 +1,0 @@
-me {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-me a {
-  color: var(--main-color);
-  text-decoration: underline;
-  transition: opacity 0.2s;
-}
-
-me a:hover {
-  opacity: 0.7;
-  transition: opacity 0.2s;
-}
-
-me main {
-  width: 100%;
-  text-align: center;
-  padding: 2rem 0;
-}
-
-me main img {
-  height: 150px;
-  margin: 1.5rem 0;
-}
-
-me main p {
-  font-weight: 378;
-  font-size: 20px;
-  line-height: 130%;
-}
-
-me main h2 {
-  font-weight: 500;
-  font-size: 36px;
-  line-height: 120%;
-  text-transform: uppercase;
-}
-
-me .faq {
-  width: 100%;
-  color: white;
-  background-color: black;
-  padding: 2rem 0;
-}
-
-me .button-holder {
-  display: flex;
-  justify-content: center;
-  margin: 1.5rem 0;
-}
-
-me main {
-  color: var(--main-color);
-  background-color: var(--main-bg-color);
-}
-
-me main h2 {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-me main .hidden {
-  display: none;
-}
-
-me main .code {
-  font-size: 0.75em;
-}
-
-me main .clipboard,
-me main .copied {
-  margin-left: 0.2em;
-  margin-top: 0.05em;
-  width: 0.75em;
-}
-
-me main .clipboard {
-  opacity: 0.5;
-  cursor: pointer;
-  transition: opacity 0.2s;
-}
-
-me main .clipboard:hover {
-  opacity: 1;
-  transition: opacity 0.2s;
-}
-
-me main .copied {
-  opacity: 0.5;
-}
-
-@media screen and (min-width: 768px) {
-
-  me main .clipboard,
-  me main .copied {
-    margin-left: 0.2em;
-    margin-top: 0.2em;
-    width: 1em;
-  }
-
-
-  me main .code {
-    font-size: 1em;
-  }
-}
